requirements.txt
langgraph
langchain-openai
discord.py
chromadb
httpx
beautifulsoup4
python-dotenv
feedparser
requests
